requirements.txt
anthropic>=0.39.0,<1.0
google-genai>=1.0.0,<2.0
Pillow>=10.0.0,<12.0
python-dotenv>=1.0.0,<2.0
fastapi>=0.115.0,<1.0
uvicorn>=0.32.0,<1.0
